remove references to window
Merge all sync components into single file

Symlink html5 sync.js into add-on
