+++ site/jsapi/include.html
@@ -282,9 +282,7 @@ function displayAppChooser(requestor, serviceName, apps, onsuccess, onerror)
 <script src="repo.js"></script>
 
 <script src="trusted.js"></script>
-<script src="sync/sync-server.js"></script>
-<script src="sync/sync-service.js"></script>
-<script src="sync/sync-scheduler.js"></script>
+<script src="sync/sync.js"></script>
 <script src="syncbuttonbridge.js"></script>
 
 </body>
