+++ sync/tests/mock-server.html
@@ -4,7 +4,7 @@
 <link rel="stylesheet" href="../../tests/doctestjs/doctest.css">
 <script src="../../tests/doctestjs/doctest.js"></script>
 <script src="../mock-server.js"></script>
-<script src="../sync-server.js"></script>
+<script src="../sync.js"></script>
 </head>
 <body class="autodoctest">
