+++ sync/tests/sync-service.html
@@ -4,8 +4,7 @@
 <link rel="stylesheet" href="../../tests/doctestjs/doctest.css">
 <script src="../../tests/doctestjs/doctest.js"></script>
 <script src="../mock-server.js"></script>
-<script src="../sync-server.js"></script>
-<script src="../sync-service.js"></script>
+<script src="../sync.js"></script>
 <script src="../../jsapi/eventmixin.js"></script>
 <script src="../../jsapi/typed-storage.js"></script>
